@@ -1,34 +1,35 @@
 buildscript {
+    ext.versions = [
+        'minSdk'             : 16,
+        'targetSdk'          : 29,
+        'compileSdk'         : 29,
+        'sourceCompatibility': JavaVersion.VERSION_1_7,
+        'targetCompatibility': JavaVersion.VERSION_1_7
+    ]
+
+    ext.deps = [
+        // Library dependencies
+        recyclerview     : 'androidx.recyclerview:recyclerview:1.1.0',
+
+        // Sample project dependencies
+        appcompat        : 'androidx.appcompat:appcompat:1.2.0',
+        material         : 'com.google.android.material:material:1.2.0',
+        smoothprogressbar: 'com.github.castorflex.smoothprogressbar:library-circular:1.1.0'
+    ]
+
     repositories {
         jcenter()
+        google()
     }
+
     dependencies {
-        classpath 'com.android.tools.build:gradle:1.5.0'
+        classpath 'com.android.tools.build:gradle:4.0.1'
     }
 }
 
 allprojects {
     repositories {
         jcenter()
+        google()
     }
 }
-
-ext {
-    minSdkVersion = 8
-    sampleMinSdkVersion = 15
-    targetSdkVersion = 23
-    compileSdkVersion = 23
-    buildToolsVersion = '23.0.1'
-}
-
-ext.deps = [
-        // Android
-        recyclerview         : 'com.android.support:recyclerview-v7:23.0.1',
-        supportv4            : 'com.android.support:support-v4:23.1.0',
-        appcompat            : 'com.android.support:appcompat-v7:23.1.0',
-        cardview             : 'com.android.support:cardview-v7:23.1.0',
-
-        // External
-        recyclerviewanimators: 'jp.wasabeef:recyclerview-animators:1.3.0',
-        smoothprogressbar    : 'com.github.castorflex.smoothprogressbar:library-circular:1.1.0'
-]
