@@ -1,9 +1,13 @@
 package com.paginate.recycler;
 
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.StaggeredGridLayoutManager;
+import android.os.Handler;
+import android.view.View;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.StaggeredGridLayoutManager;
 
 import com.paginate.Paginate;
 
@@ -15,12 +19,14 @@ public final class RecyclerPaginate extends Paginate {
     private WrapperAdapter wrapperAdapter;
     private WrapperSpanSizeLookup wrapperSpanSizeLookup;
 
-    RecyclerPaginate(RecyclerView recyclerView,
-                     Paginate.Callbacks callbacks,
-                     int loadingTriggerThreshold,
-                     boolean addLoadingListItem,
-                     LoadingListItemCreator loadingListItemCreator,
-                     LoadingListItemSpanLookup loadingListItemSpanLookup) {
+    RecyclerPaginate(
+        RecyclerView recyclerView,
+        Paginate.Callbacks callbacks,
+        int loadingTriggerThreshold,
+        boolean addLoadingListItem,
+        LoadingListItemCreator loadingListItemCreator,
+        LoadingListItemSpanLookup loadingListItemSpanLookup
+    ) {
         this.recyclerView = recyclerView;
         this.callbacks = callbacks;
         this.loadingTriggerThreshold = loadingTriggerThreshold;
@@ -29,18 +35,21 @@ public final class RecyclerPaginate extends Paginate {
         recyclerView.addOnScrollListener(mOnScrollListener);
 
         if (addLoadingListItem) {
-            // Wrap existing adapter with new adapter that will add loading row
+            // Wrap existing adapter with new WrapperAdapter that will add loading row
             RecyclerView.Adapter adapter = recyclerView.getAdapter();
+            assert adapter != null;
             wrapperAdapter = new WrapperAdapter(adapter, loadingListItemCreator);
+            wrapperAdapter.displayLoadingRow(!callbacks.hasLoadedAllItems());
             adapter.registerAdapterDataObserver(mDataObserver);
             recyclerView.setAdapter(wrapperAdapter);
 
             // For GridLayoutManager use separate/customisable span lookup for loading row
             if (recyclerView.getLayoutManager() instanceof GridLayoutManager) {
                 wrapperSpanSizeLookup = new WrapperSpanSizeLookup(
-                        ((GridLayoutManager) recyclerView.getLayoutManager()).getSpanSizeLookup(),
-                        loadingListItemSpanLookup,
-                        wrapperAdapter);
+                    ((GridLayoutManager) recyclerView.getLayoutManager()).getSpanSizeLookup(),
+                    loadingListItemSpanLookup,
+                    wrapperAdapter
+                );
                 ((GridLayoutManager) recyclerView.getLayoutManager()).setSpanSizeLookup(wrapperSpanSizeLookup);
             }
         }
@@ -62,9 +71,9 @@ public final class RecyclerPaginate extends Paginate {
         recyclerView.removeOnScrollListener(mOnScrollListener);   // Remove scroll listener
         if (recyclerView.getAdapter() instanceof WrapperAdapter) {
             WrapperAdapter wrapperAdapter = (WrapperAdapter) recyclerView.getAdapter();
-            RecyclerView.Adapter adapter = wrapperAdapter.getWrappedAdapter();
-            adapter.unregisterAdapterDataObserver(mDataObserver); // Remove data observer
-            recyclerView.setAdapter(adapter);                     // Swap back original adapter
+            RecyclerView.Adapter originalAdapter = wrapperAdapter.getWrappedAdapter();
+            originalAdapter.unregisterAdapterDataObserver(mDataObserver); // Remove data observer
+            swapBackAdapter(originalAdapter);                             // Swap back original adapter
         }
         if (recyclerView.getLayoutManager() instanceof GridLayoutManager && wrapperSpanSizeLookup != null) {
             // Swap back original SpanSizeLookup
@@ -73,7 +82,46 @@ public final class RecyclerPaginate extends Paginate {
         }
     }
 
-    void checkEndOffset() {
+    private void swapBackAdapter(final RecyclerView.Adapter adapter) {
+        final RecyclerView.LayoutManager layoutManager = recyclerView.getLayoutManager();
+        if (layoutManager instanceof LinearLayoutManager) {
+            LinearLayoutManager linearManager = (LinearLayoutManager) layoutManager;
+
+            // get first visible view and its position
+            int firstVisibleItemPosition = linearManager.findFirstVisibleItemPosition();
+            View firstVisibleChild = layoutManager.findViewByPosition(firstVisibleItemPosition);
+
+            // get scroll offsets of the view
+            int xPixelScroll = 0;
+            int yPixelScroll = 0;
+            if (firstVisibleChild != null) {
+                if (!linearManager.getReverseLayout()) {
+                    xPixelScroll = firstVisibleChild.getLeft();
+                    yPixelScroll = firstVisibleChild.getTop();
+                } else {
+                    xPixelScroll = firstVisibleChild.getRight() - recyclerView.getWidth();
+                    yPixelScroll = firstVisibleChild.getBottom() - recyclerView.getHeight();
+                }
+            }
+
+            // set adapter and scroll to the item position that was displayed
+            // when WrapperAdapter was used
+            recyclerView.setAdapter(adapter);
+            recyclerView.scrollToPosition(firstVisibleItemPosition);
+            recyclerView.scrollBy(-xPixelScroll, -yPixelScroll);
+        } else if (layoutManager instanceof StaggeredGridLayoutManager) {
+            recyclerView.setAdapter(adapter);
+            // https://issuetracker.google.com/issues/37017287
+            new Handler().post(new Runnable() {
+                @Override
+                public void run() {
+                    ((StaggeredGridLayoutManager) layoutManager).invalidateSpanAssignments();
+                }
+            });
+        }
+    }
+
+    private void checkEndOffset() {
         int visibleItemCount = recyclerView.getChildCount();
         int totalItemCount = recyclerView.getLayoutManager().getItemCount();
 
@@ -92,8 +140,7 @@ public final class RecyclerPaginate extends Paginate {
         }
 
         // Check if end of the list is reached (counting threshold) or if there is no items at all
-        if ((totalItemCount - visibleItemCount) <= (firstVisibleItemPosition + loadingTriggerThreshold)
-                || totalItemCount == 0) {
+        if ((totalItemCount - visibleItemCount) <= (firstVisibleItemPosition + loadingTriggerThreshold) || totalItemCount == 0) {
             // Call load more only if loading is not currently in progress and if there is more items to load
             if (!callbacks.isLoading() && !callbacks.hasLoadedAllItems()) {
                 callbacks.onLoadMore();
@@ -108,7 +155,7 @@ public final class RecyclerPaginate extends Paginate {
 
     private final RecyclerView.OnScrollListener mOnScrollListener = new RecyclerView.OnScrollListener() {
         @Override
-        public void onScrolled(RecyclerView recyclerView, int dx, int dy) {
+        public void onScrolled(@NonNull RecyclerView recyclerView, int dx, int dy) {
             checkEndOffset(); // Each time when list is scrolled check if end of the list is reached
         }
     };
@@ -167,11 +214,11 @@ public final class RecyclerPaginate extends Paginate {
         }
 
         /**
-         * Set the offset from the end of the list at which the load more event needs to be triggered. Default offset
-         * if 5.
+         * Set the offset from the end of the list at which the load more event needs to be triggered.
+         * Default offset if 5.
          *
          * @param threshold number of items from the end of the list.
-         * @return {@link com.paginate.recycler.RecyclerPaginate.Builder}
+         * @return {@link RecyclerPaginate.Builder}
          */
         public Builder setLoadingTriggerThreshold(int threshold) {
             this.loadingTriggerThreshold = threshold;
@@ -180,14 +227,14 @@ public final class RecyclerPaginate extends Paginate {
 
         /**
          * Setup loading row. If loading row is used original adapter set on RecyclerView will be wrapped with
-         * internal adapter that will add loading row as the last item in the list. Paginate will observer the
+         * internal adapter that will add loading row as the last item in the list. Paginate will observe the
          * changes upon original adapter and remove loading row if there is no more data to load. By default loading
          * row will be added.
          *
          * @param addLoadingListItem true if loading row needs to be added, false otherwise.
-         * @return {@link com.paginate.recycler.RecyclerPaginate.Builder}
-         * @see {@link com.paginate.Paginate.Callbacks#hasLoadedAllItems()}
-         * @see {@link com.paginate.recycler.RecyclerPaginate.Builder#setLoadingListItemCreator(LoadingListItemCreator)}
+         * @return {@link RecyclerPaginate.Builder}
+         * @see Paginate.Callbacks#hasLoadedAllItems()
+         * @see RecyclerPaginate.Builder#setLoadingListItemCreator(LoadingListItemCreator)
          */
         public Builder addLoadingListItem(boolean addLoadingListItem) {
             this.addLoadingListItem = addLoadingListItem;
@@ -198,7 +245,7 @@ public final class RecyclerPaginate extends Paginate {
          * Set custom loading list item creator. If no creator is set default one will be used.
          *
          * @param creator Creator that will ne called for inflating and binding loading list item.
-         * @return {@link com.paginate.recycler.RecyclerPaginate.Builder}
+         * @return {@link RecyclerPaginate.Builder}
          */
         public Builder setLoadingListItemCreator(LoadingListItemCreator creator) {
             this.loadingListItemCreator = creator;
@@ -209,8 +256,8 @@ public final class RecyclerPaginate extends Paginate {
          * Set custom SpanSizeLookup for loading list item. Use this when {@link GridLayoutManager} is used and
          * loading list item needs to have custom span. Full span of {@link GridLayoutManager} is used by default.
          *
-         * @param loadingListItemSpanLookup LoadingListItemSpanLookup that will be called for loading list item span.
-         * @return {@link com.paginate.recycler.RecyclerPaginate.Builder}
+         * @param loadingListItemSpanLookup {@link LoadingListItemSpanLookup} that will be called for getting list item span.
+         * @return {@link RecyclerPaginate.Builder}
          */
         public Builder setLoadingListItemSpanSizeLookup(LoadingListItemSpanLookup loadingListItemSpanLookup) {
             this.loadingListItemSpanLookup = loadingListItemSpanLookup;
@@ -226,6 +273,7 @@ public final class RecyclerPaginate extends Paginate {
             if (recyclerView.getAdapter() == null) {
                 throw new IllegalStateException("Adapter needs to be set!");
             }
+
             if (recyclerView.getLayoutManager() == null) {
                 throw new IllegalStateException("LayoutManager needs to be set on the RecyclerView");
             }
@@ -238,9 +286,8 @@ public final class RecyclerPaginate extends Paginate {
                 loadingListItemSpanLookup = new DefaultLoadingListItemSpanLookup(recyclerView.getLayoutManager());
             }
 
-            return new RecyclerPaginate(recyclerView, callbacks, loadingTriggerThreshold, addLoadingListItem,
-                    loadingListItemCreator, loadingListItemSpanLookup);
+            return new RecyclerPaginate(recyclerView, callbacks, loadingTriggerThreshold,
+                addLoadingListItem, loadingListItemCreator, loadingListItemSpanLookup);
         }
     }
-
 }
