@@ -1,18 +1,24 @@
 package com.example.paginate.adapter;
 
-import android.support.v7.widget.RecyclerView;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.example.paginate.R;
 import com.example.paginate.data.Person;
 
 import java.util.List;
+import java.util.Locale;
 
 public class RecyclerPersonAdapter extends RecyclerView.Adapter<RecyclerPersonAdapter.PersonVH> implements RecyclerOnItemClickListener {
+
+    private static final String TAG = "RecyclerPersonAdapter";
 
     private final List<Person> data;
 
@@ -20,6 +26,7 @@ public class RecyclerPersonAdapter extends RecyclerView.Adapter<RecyclerPersonAd
         this.data = data;
     }
 
+    @NonNull
     @Override
     public PersonVH onCreateViewHolder(ViewGroup parent, int viewType) {
         View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.person_list_item, parent, false);
@@ -29,12 +36,26 @@ public class RecyclerPersonAdapter extends RecyclerView.Adapter<RecyclerPersonAd
     @Override
     public void onBindViewHolder(PersonVH holder, final int position) {
         Person person = data.get(position);
-        holder.tvFullName.setText(String.format("%s %s, %d", person.getFirstName(), person.getLastName(), person.getAge()));
+        holder.tvFullName.setText(
+            String.format(Locale.getDefault(), "%s %s, %d", person.getFirstName(), person.getLastName(), person.getAge())
+        );
     }
 
     @Override
     public int getItemCount() {
         return data.size();
+    }
+
+    @Override
+    public void onAttachedToRecyclerView(@NonNull RecyclerView recyclerView) {
+        super.onAttachedToRecyclerView(recyclerView);
+        Log.d(TAG, "onAttachedToRecyclerView");
+    }
+
+    @Override
+    public void onDetachedFromRecyclerView(@NonNull RecyclerView recyclerView) {
+        super.onDetachedFromRecyclerView(recyclerView);
+        Log.d(TAG, "onDetachedFromRecyclerView");
     }
 
     @Override
@@ -45,9 +66,14 @@ public class RecyclerPersonAdapter extends RecyclerView.Adapter<RecyclerPersonAd
     }
 
     public void add(List<Person> items) {
-        int previousDataSize = this.data.size();
-        this.data.addAll(items);
-        notifyItemRangeInserted(previousDataSize, items.size());
+        data.addAll(items);
+        notifyDataSetChanged();
+    }
+
+    public void replace(List<Person> items) {
+        data.clear();
+        data.addAll(items);
+        notifyDataSetChanged();
     }
 
     public static class PersonVH extends RecyclerView.ViewHolder {
